Move TLS test to test_ldapclient, add TLS timeout test
