+++ tests/test_ldapconnection.py
@@ -287,20 +287,6 @@ class LDAPConnectionTest(unittest.TestCase):
                         self.cfg["SIMPLEAUTH"]["adusername"]]
         self.assertIn(obj, expected_res)
 
-    def test_tls(self):
-        """ Test TLS connection. """
-        if self.cfg['SERVER']['has_tls'] == 'False':
-            self.skipTest("TLS is not set.")
-        client = LDAPClient(self.url, True)
-        client.set_cert_policy("ALLOW")
-        client.set_ca_cert(None)
-        client.set_ca_cert_dir(None)
-        try:
-            conn = client.connect()
-            conn.close()
-        except Exception as exc:
-            self.fail("TLS connection is failed with: %s" % str(exc))
-
     def test_connection_error(self):
         """ Test connection error. """
         client = LDAPClient("ldap://invalid")
